Add custom options to Ajv and compile schemas
This commit introduces support for custom options passed to the Ajv validation library for more flexible schema validation. It also changes our method of storing schemas to pre-compile them, which will improve the efficiency of the validation process. From now on, the schema map will contain compiled ValidateFunctions instead of raw schemas that need to be compiled every time. This should speed up validation checks as the schema doesn't need to be compiled for each validation check.