Add HTTP publish interface to http tests and verify it works
The update introduces an HTTP interface for publishing to the LogStore plugin. It also extends the test suite to include new scenarios that checks if publishing through the HTTP interface properly stores entries on LogStore. Further, it includes a permission model to allow the LogStore node to publish on a specific stream.