Add tinyproxy installation script and integration
A new script to automate the installation of tinyproxy has been added, which checks for existing installation, if the script is being run with proper privileges and the presence of either apt-get or yum package managers. If everything checks out, the tinyproxy gets installed and its installation is verified. This script gets triggered after the package installation (postinstall). Integrations in the package core and Dockerfile also updated to reflect these changes. Error handling for tinyproxy startup also added in tinyProxy.ts.