Update tinyproxy installation script and enhance project paths
The tinyproxy installation in install-tinyscript.sh has been updated to download the binary directly from GitHub, instead of relying on apt-get or yum package managers. A new utility file paths.ts is added in project to provide the root path of the project in a more reliable way. Also modified Dockerfile and http-proxy constants file according to these changes.