Add "includeOnly" field to core configuration
The "includeOnly" field has been added to the core configuration and schema. This is an optional array that takes in glob-like patterns specifying the streams that should be tracked by the node. It's used as a filter; if the list is empty, all streams fetched from the contract will be tracked.