Add stream filtering based on glob patterns
The update introduces glob-based stream filtering in the `LogStoreNetworkConfig` and `LogStoreNetworkPlugin`. This was achieved by adding 'micromatch' to the dependencies and creating a utility function 'globsMatch' that checks if a string matches any provided glob patterns. This function is now used to filter streams based on the newly added 'includeOnly' option in the plugins.