Refactor filter logic for stream inclusion
Refactored the filter logic for stream inclusion from the streamFilter() function to a newly introduced isStreamIncluded() function, improving code reusability. In addition, we introduced the shouldPublishQueryResponseFn() function in order to decide whether the query response should be published if the node opts to not participate in the propagation of the query response.