Add tests and utilities for log store node
The commit introduces utilities needed for effective management of 'logStoreNode' - start, stop, setup, and tear down of nodes. There's also enhancement on the utilities for account setup and cleanup. Alongside the utility scripts, a new integration test was added to verify if queries from a partially included stream are handled correctly by the 'logStoreNode'.