Add queryFirst method to LogStore and DatabaseAdapter
This update introduces the queryFirst method to both the LogStore and DatabaseAdapter classes. QueryFirst fetches a specified number of records from the beginning of the log, contrary to queryLast which fetches records from the end. Error checking was also added to ensure 'requestCount' is handled correctly when it is a negative value.