Update Docker build and push process in GitHub workflow
This commit changes the way the Docker image is built and pushed in the GitHub workflow. The code to manually build the Docker image has been commented out, and a new 'build-push-action' is now used. This action simplifies the process by building and pushing the image in a single step. Additionally, the Docker image tagging section has been modified to streamline this process.