Update Docker image tagging in GitHub Actions workflow
Modified the 'deploy-image.yml' workflow in the GitHub Actions to change how Docker images are tagged. The image name is now directly dependent on the 'env.REGISTRY' path instead of a custom image name variable, and the tagging structure is consequently altered. This ensures tags streamline with branch types and either labels latest images accordingly or uses the branch name.