Add case insensitivity to owner address in http tests
In the http tests, a new condition has been added to ensure that the owner's address is case insensitive. The validation was done by performing queries using different cases of the owner address and asserting that all responses are equal. This increases the robustness of the tests to handle various cases of addresses.